backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary,pool]==3.1.13
pydantic>=2.4.0,<3.0.0
python-multipart==0.0.6
langchain>=0.2.12
langgraph>=0.0.40
openai>=1.10.0
langchain-openai>=0.1.0
google-generativeai>=0.7.2
langchain-google-genai>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0